ACB-8: Rename the bundle and prepare to akeneo-labs
