+++ Denormalizer/ProductRule/SetCategoryActionDenormalizer.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace PimEnterprise\Bundle\AutomaticClassificationBundle\Denormalizer\ProductRule;
+namespace PimEnterprise\Bundle\ClassificationRuleBundle\Denormalizer\ProductRule;
 
-use PimEnterprise\Bundle\AutomaticClassificationBundle\Model\ProductSetCategoryActionInterface;
+use PimEnterprise\Bundle\ClassificationRuleBundle\Model\ProductSetCategoryActionInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 
 /**
