+++ DependencyInjection/PimEnterpriseClassificationRuleExtension.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PimEnterprise\Bundle\AutomaticClassificationBundle\DependencyInjection;
+namespace PimEnterprise\Bundle\ClassificationRuleBundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -14,7 +14,7 @@ use Symfony\Component\HttpKernel\DependencyInjection\Extension;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PimEnterpriseAutomaticClassificationExtension extends Extension
+class PimEnterpriseClassificationRuleExtension extends Extension
 {
     /**
      * {@inheritdoc}
